+++ src/encoder.rs
@@ -1,7 +1,7 @@
 //! # Encoding utilities
-//! 
+//!
 //! Some simple utilities to encode an integer to string
-//! 
+//!
 
 pub const BASE_62: &str = "0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ";
 pub const BASE_63: &str = "0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ_";