+++ src/namer.rs
@@ -1,5 +1,5 @@
 //! # Naming utilities
-//! 
+//!
 //! Module dedicated to naming of objects such as Qrlew expressions and relations
 //!